feat: component move and impl styles
